Fixed unresolved resource dependencies (#4)
Something has changed over the year in AWS CloudFormation and the
logic that was conditionally creating even listeners depending on
whether we are working with the selected pipelines and projects
or with all of them was no longer performing as expected.

Instead of investigating why it became broken, I used it as an
opportunity to optimise the template a bit and bring it to the
current convention for all my YAML templates.  This reduced the
size of the template as well since I got rid of repetitive
resource definitions.