Updated README.md to reflect AWS changes for S3
Adjusted information about the TemplateURL for the nested stack
examples (fixes #2).

Additionally to this added a section that handles the recent AWS
change (Nov'18) when the generated badges are not publicly
accessable if the new feature of the S3 service is leveraged.
Provided a couple of options on how to solve the issue.

Fixed some minor typos in the file.